@LuisPH3 2024-12-05
> StaticOverhead in isolation, the slowdown is less significant.

Does this mean that a re-run of StaticOverhead did not have the mentioned slowdown? or less of it?
Maybe it is time to study the behavior of the benchmark engine: order, effects on cache pollution/preload, and garbage collector.