@LorenzSchueler 2024-12-05
because addition and subtraction are easy to implement. For multiplication we need a result type that has twice as many bytes (u64*u64->u128) so this will not work because there is no u256 in the std lib.